AILearningPlatform/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openai>=1.78.0",
    "psycopg2-binary>=2.9.10",
    "werkzeug>=3.1.3",
    "sqlalchemy>=2.0.40",
    "pyjwt>=2.10.1",
    "flask-dance>=7.1.0",
    "oauthlib>=3.2.2",
    "Pillow>=10.0.0",  
    "requests>=2.30.0", 
    "python-dateutil>=2.8.0",
    "pandas>=2.0.0", # Added for CSV handling in chatbot
    "Flask-WTF>=1.2.0" # Added for CSRF protection
]
